parwave.c: the noise parameter for gen_noise is not used as input.
The noise variable is used in gen_noise, but not using the value
passed to it. It is used as a local variable.

As such, this change removes the noise parameter and reworks
gen_noise to set the nlast variable directly.

This also avoids potential confusion of the use of gen_noise
where the variable can be seen as uninitialised and that it should
be static. Both of these are not the case as can be seen above.